:recycle: Refactor imports to use new UI structure and rename files for consistency
